+++ src/fcst/automation/__init__.py
@@ -5,9 +5,6 @@ automation sub-package
 Automation, make it easy for common forecasting tasks.
 """
 
-from ._forecasting_automation import (
-    run_forecasting_automation,
-    run_forecasting_pipeline,
-)
+from ._forecasting_automation import run_forecasting_automation
 
-__all__ = ["run_forecasting_automation", "run_forecasting_pipeline"]
+__all__ = ["run_forecasting_automation"]