+++ src/fcst/automation/_forecasting_automation.py
@@ -1,4 +1,5 @@
 import warnings
+from collections.abc import Iterable
 from typing import Tuple, overload
 
 import pandas as pd
@@ -9,43 +10,40 @@ from ..evaluation.model_evaluation import backtest_evaluate
 from ..evaluation.model_selection import select_best_models
 from ..forecasting.ensemble import ensemble_forecast
 from ..models import base_models
-from ..preprocessing.timeseries import extract_timeseries
+from ..preprocessing import prepare_timeseries
 
 
 @overload
-def run_forecasting_pipeline(
+def _forecasting_pipeline(
     series: pd.Series,
     backtest_periods: int,
     eval_periods: int,
     top_n: int,
     forecasting_periods: int,
     models: ModelDict = base_models,
-    id_: str = "",
     return_backtest_results: bool = False,
-) -> Tuple[str, pd.DataFrame]: ...
+) -> pd.DataFrame: ...
 
 
 @overload
-def run_forecasting_pipeline(
+def _forecasting_pipeline(
     series: pd.Series,
     backtest_periods: int,
     eval_periods: int,
     top_n: int,
     forecasting_periods: int,
     models: ModelDict = base_models,
-    id_: str = "",
     return_backtest_results: bool = True,
-) -> Tuple[str, pd.DataFrame, pd.DataFrame]: ...
+) -> Tuple[pd.DataFrame, pd.DataFrame]: ...
 
 
-def run_forecasting_pipeline(
+def _forecasting_pipeline(
     series: pd.Series,
     backtest_periods: int,
     eval_periods: int,
     top_n: int,
     forecasting_periods: int,
     models: ModelDict = base_models,
-    id_: str = "",
     return_backtest_results: bool = False,
 ) -> Tuple[str, pd.DataFrame]:
     """Performs model selection and ensemble forecast for a single time-series
@@ -67,73 +65,65 @@ def run_forecasting_pipeline(
 
         models: (ModelDict): A dictionary of models to use in forecasting (Default = base_models)
 
-        id_ (str): ID identifying the series (Default = "")
-
         return_backtest_results (bool): Whether or not to return the back-testing raw results (Default is False),
 
     Returns
     -------
         Tuple[str, pd.Series]: ID and the resulting forecasted series (when return_backtest_results = False)
 
-        Tuple[str, pd.Series, pd.DataFrame]: ID and the resulting forecasted series with the back-testing raw results (when return_backtest_results = True)
+        Tuple[pd.Series, pd.DataFrame]: ID and the resulting forecasted series with the back-testing raw results (when return_backtest_results = True)
     """
+    with warnings.catch_warnings():
+        # Suppress all warnings from inside this function
+        warnings.simplefilter("ignore")
 
-    try:
-        model_results = backtest_evaluate(
-            series,
-            models,
-            backtest_periods=backtest_periods,
-            eval_periods=eval_periods,
-            return_results=return_backtest_results,
-        )
+        try:
+            model_results = backtest_evaluate(
+                series,
+                models,
+                backtest_periods=backtest_periods,
+                eval_periods=eval_periods,
+                return_results=return_backtest_results,
+            )
 
-        if return_backtest_results:
-            model_results, df_backtest_results = model_results[0], model_results[1]
+            if return_backtest_results:
+                model_results, df_backtest_results = model_results[0], model_results[1]
 
-        models_list = select_best_models(model_results=model_results, top_n=top_n)
+            models_list = select_best_models(model_results=model_results, top_n=top_n)
 
-        forecast_results = ensemble_forecast(
-            models=models,
-            model_names=models_list,
-            series=series,
-            periods=forecasting_periods,
-        )
+            forecast_results = ensemble_forecast(
+                models=models,
+                model_names=models_list,
+                series=series,
+                periods=forecasting_periods,
+            )
 
-        df_forecast_results = pd.DataFrame(forecast_results)
-        df_forecast_results["selected_models"] = "|".join(models_list)
+            df_forecast_results = pd.DataFrame(forecast_results)
+            df_forecast_results["selected_models"] = "|".join(models_list)
 
-        if return_backtest_results:
-            return id_, df_forecast_results, df_backtest_results
+            if return_backtest_results:
+                return df_forecast_results, df_backtest_results
 
-        return id_, df_forecast_results
+            return df_forecast_results
 
-    except Exception as e:
-        print("Unexpected error occurred for ID:", id_, e)
-
-
-# df_raw: pd.DataFrame,
-# date_col: str,
-# value_col: str,
-# data_period_date: pd.Period,
-# id_cols: list[str] | None = None,
-# min_cap: int | None = 0,
-# freq: str = "M",
-# id_join_char: str = "_",
-
-# TODO: Change forecasting automation to run from df_raw
-# TODO: Remove forecasting pipeline as df_raw can be series
+        except Exception as e:
+            print("Unexpected error occurred for ID:", e)
 
 
 @overload
 def run_forecasting_automation(
-    df_forecasting: pd.DataFrame,
+    df_raw: pd.DataFrame,
+    date_col: str,
     value_col: str,
     data_period_date: pd.Period,
     backtest_periods: int,
     eval_periods: int,
     top_n: int,
     forecasting_periods: int,
-    id_col: str = "id",
+    id_cols: list[str] | None = None,
+    id_join_char: str = "_",
+    min_cap: int | None = 0,
+    freq: str = "M",
     models: ModelDict = base_models,
     return_backtest_results: bool = False,
     parallel: bool = True,
@@ -143,30 +133,38 @@ def run_forecasting_automation(
 
 @overload
 def run_forecasting_automation(
-    df_forecasting: pd.DataFrame,
+    df_raw: pd.DataFrame,
+    date_col: str,
     value_col: str,
     data_period_date: pd.Period,
     backtest_periods: int,
     eval_periods: int,
     top_n: int,
     forecasting_periods: int,
-    id_col: str = "id",
+    id_cols: list[str] | None = None,
+    id_join_char: str = "_",
+    min_cap: int | None = 0,
+    freq: str = "M",
     models: ModelDict = base_models,
-    return_backtest_results: bool = False,
+    return_backtest_results: bool = True,
     parallel: bool = True,
     n_jobs: int = -1,
 ) -> Tuple[pd.DataFrame, pd.DataFrame]: ...
 
 
 def run_forecasting_automation(
-    df_forecasting: pd.DataFrame,
+    df_raw: pd.DataFrame,
+    date_col: str,
     value_col: str,
     data_period_date: pd.Period,
     backtest_periods: int,
     eval_periods: int,
     top_n: int,
     forecasting_periods: int,
-    id_col: str = "id",
+    id_cols: list[str] | None = None,
+    id_join_char: str = "_",
+    min_cap: int | None = 0,
+    freq: str = "M",
     models: ModelDict = base_models,
     return_backtest_results: bool = False,
     parallel: bool = True,
@@ -186,10 +184,9 @@ def run_forecasting_automation(
 
     Parameters
     ----------
-        df_forecasting (pd.DataFrame): Preprocessed DF for forecasting
-            Where the index is the pd.PeriodIndex,
-            and the columns are id and value.
-            The values are resampled to the specified `freq`.
+        df_raw (pd.DataFrame): Raw DF that has a date column, other info, and the value to forecast
+
+        date_col (str): The date column to use in forecasting
 
         value_col (str): Column to forecast
 
@@ -203,7 +200,18 @@ def run_forecasting_automation(
 
         forecasting_periods (int): Forecasting periods
 
-        id_col (str): ID column name to distinguish between series
+        id_cols (list[str] | None): A list containing the column names to create a unique time-series ID (Default is None)
+            If None, the whole dataframe is treated as a single time-series
+            If a list of columns is poassed in, a new "id" index will be created
+
+        id_join_char (str): A character to join multiple ID columns (Default = "_")
+
+        min_cap (int | None): Minimum value to cap before forecast
+            If set, the value is used to set the minimum.
+            For example, you might want to set 0 for sales.
+            If None, use the existing values.
+
+        freq (str): Frequency to resample and forecast (Default = "M")
 
         models: (ModelDict): A dictionary of models to use in forecasting (Default = base_models)
 
@@ -222,38 +230,49 @@ def run_forecasting_automation(
 
     models = models.copy()
 
-    def _fcst(id_, series):  # Internal function for delayed, parallel
+    def _fcst(series):  # Internal function for simplicity
         with warnings.catch_warnings():
             # Suppress all warnings from inside this function
             warnings.simplefilter("ignore")
-            return run_forecasting_pipeline(
+            return _forecasting_pipeline(
                 series=series,
                 backtest_periods=backtest_periods,  # Constant
                 eval_periods=eval_periods,  # Constant
                 top_n=top_n,  # Constant
                 forecasting_periods=forecasting_periods,  # Constant
                 models=models,  # Constant
-                id_=id_,
                 return_backtest_results=return_backtest_results,
             )
 
-    each_series = extract_timeseries(
-        df_forecasting,
+    timeseries: pd.Series | Iterable[Tuple[str, pd.Series]] = prepare_timeseries(
+        df_raw=df_raw,
+        date_col=date_col,
         value_col=value_col,
         data_period_date=data_period_date,
-        id_col=id_col,
+        id_cols=id_cols,
+        min_cap=min_cap,
+        freq=freq,
+        id_join_char=id_join_char,
     )
 
-    # Run in parallel
+    if not id_cols:
+        return _fcst(series=timeseries)
+
+    # Check if run in parallel
     if parallel:
-        results = Parallel(n_jobs=n_jobs)(
-            delayed(_fcst)(id_, series) for id_, series in each_series
+        timeseries_list = list(timeseries)
+        results_list = Parallel(n_jobs=n_jobs)(
+            (delayed(_fcst)(series)) for _, series in timeseries_list
         )
+
+        results = [
+            (id_, result) for (id_, _), result in zip(timeseries_list, results_list)
+        ]
     else:
-        results = [_fcst(id_, series) for id_, series in each_series]
+        results = [(id_, _fcst(series)) for id_, series in timeseries]
 
     def _filter_none_results(results_list: list[Tuple[str, pd.Series]]):
-        return list(filter(lambda x: x is not None, results_list))
+        return list(filter(lambda x: x[1] is not None, results_list))
 
     def _get_df_forecasting_from_each_result(result: Tuple[str, pd.Series]):
         id_ = result[0]